Use double for coordinates in spatial APIs and DTOs
Update all spatial search APIs, DTOs, mappers, services, and tests to use double instead of decimal for latitude, longitude, and related fields. This improves compatibility and precision for spatial calculations. All method signatures, geometry creation, and validation logic are updated accordingly. No business logic changes; only data types are affected.

+++ PreflightApi.API/Controllers/AirportController.cs
@@ -81,8 +81,8 @@ public class AirportController(
     [ProducesResponseType(typeof(PaginatedResponse<AirportDto>), StatusCodes.Status200OK)]
     [ProducesResponseType(typeof(ApiErrorResponse), StatusCodes.Status400BadRequest)]
     public async Task<ActionResult<PaginatedResponse<AirportDto>>> SearchNearby(
-        [FromQuery] decimal lat,
-        [FromQuery] decimal lon,
+        [FromQuery] double lat,
+        [FromQuery] double lon,
         [FromQuery] PaginationParams pagination,
         CancellationToken ct,
         [FromQuery] double radiusNm = 30)
